requirements2.txt
async-timeout==4.0.3
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
redis==5.0.4
requests==2.31.0
simple-websocket==1.0.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.18.1
zope.event==5.0
zope.interface==6.4.post2
